+++ pkg/dataset/duration.go
@@ -1,9 +1,15 @@
+// This file and its contents are licensed under the Apache License 2.0.
+// Please see the included NOTICE for copyright information and
+// LICENSE for a copy of the license.
 package dataset
 
 import (
 	"fmt"
+	"reflect"
 	"strings"
 	"time"
+
+	"github.com/mitchellh/mapstructure"
 )
 
 const (
@@ -63,4 +69,29 @@ func handleDays(s []byte) (time.Duration, error) {
 // String returns a string value of DayDuration.
 func (d DayDuration) String() string {
 	return time.Duration(d).String()
+}
+
+// StringToDayDurationHookFunc returns a mapstructure.DecodeHookFunc that
+// converts strings to DayDuration.
+func StringToDayDurationHookFunc() mapstructure.DecodeHookFunc {
+	return func(
+		f reflect.Type,
+		t reflect.Type,
+		data interface{}) (interface{}, error) {
+		if f.Kind() != reflect.String {
+			return data, nil
+		}
+
+		var d DayDuration
+
+		if t != reflect.TypeOf(d) {
+			return data, nil
+		}
+
+		err := d.UnmarshalText([]byte(data.(string)))
+		if err != nil {
+			return nil, err
+		}
+		return DayDuration(d), nil
+	}
 }