+++ pkg/pgmodel/ingestor/trace/instrumentation_lib_batch.go
@@ -53,7 +53,7 @@ func (il instrumentationLibrary) ScanIDs(r pgx.BatchResults) (interface{}, error
 }
 
 // instrumentationLibraryBatch queues up items to send to the db but it sorts before sending
-//this avoids deadlocks in the db
+// this avoids deadlocks in the db
 type instrumentationLibraryBatch struct {
 	b batcher
 }