+++ pkg/pgmodel/querier/series_set_test.go
@@ -23,7 +23,7 @@ import (
 	"github.com/timescale/promscale/pkg/pgmodel/model"
 )
 
-//nolint
+//nolint:all
 type mockPgxRows struct {
 	closeCalled  bool
 	firstRowRead bool
@@ -32,34 +32,38 @@ type mockPgxRows struct {
 	err          error
 }
 
-//nolint
 // Close closes the rows, making the connection ready for use again. It is safe
 // to call Close after rows is already closed.
+//
+//nolint:all
 func (m *mockPgxRows) Close() {
 	m.closeCalled = true
 }
 
-//nolint
 // Err returns any error that occurred while reading.
+//
+//nolint:all
 func (m *mockPgxRows) Err() error {
 	return nil
 }
 
-//nolint
 // CommandTag returns the command tag from this query. It is only available after Rows is closed.
+//
+//nolint:all
 func (m *mockPgxRows) CommandTag() pgconn.CommandTag {
 	panic("not implemented")
 }
 
-//nolint
+//nolint:all
 func (m *mockPgxRows) FieldDescriptions() []pgproto3.FieldDescription {
 	panic("not implemented")
 }
 
-//nolint
 // Next prepares the next row for reading. It returns true if there is another
 // row and false if no more rows are available. It automatically closes rows
 // when all rows are read.
+//
+//nolint:all
 func (m *mockPgxRows) Next() bool {
 	if m.firstRowRead {
 		m.idx++
@@ -69,11 +73,12 @@ func (m *mockPgxRows) Next() bool {
 	return m.idx < len(m.results)
 }
 
-//nolint
 // Scan reads the values from the current row into dest values positionally.
 // dest can include pointers to core types, values implementing the Scanner
 // interface, []byte, and nil. []byte will skip the decoding process and directly
 // copy the raw bytes received from PostgreSQL. nil will skip the value entirely.
+//
+//nolint:all
 func (m *mockPgxRows) Scan(dest ...interface{}) error {
 	if m.err != nil {
 		return m.err
@@ -103,20 +108,22 @@ func (m *mockPgxRows) Scan(dest ...interface{}) error {
 	return nil
 }
 
-//nolint
 // Values returns the decoded row values.
+//
+//nolint:all
 func (m *mockPgxRows) Values() ([]interface{}, error) {
 	panic("not implemented")
 }
 
-//nolint
 // RawValues returns the unparsed bytes of the row values. The returned [][]byte is only valid until the next Next
 // call or the Rows is closed. However, the underlying byte data is safe to retain a reference to and mutate.
+//
+//nolint:all
 func (m *mockPgxRows) RawValues() [][]byte {
 	panic("not implemented")
 }
 
-//nolint
+//nolint:all
 func generateArrayHeader(numDim, containsNull, elemOID, arrayLength uint32, addData []byte) []byte {
 	result := make([]byte, 20)
 
@@ -460,7 +467,7 @@ func (m mapQuerier) LabelsForIdMap(idMap map[int64]labels.Label) (err error) {
 	return nil
 }
 
-//nolint
+//nolint:all
 func genRows(count int) [][][]byte {
 	result := make([][][]byte, count)