+++ CHANGELOG.md
@@ -17,6 +17,12 @@ We use the following categories for changes:
 
 ### Added
 - Alerts from promscale monitoring mixin are groupped also by namespace label [#1714]
+- Added a new family of metrics tracking database maintenance jobs durations and failures [#1745]
+- Allow config options in file to be set as yaml mappings [#1737]
+- Add `startup.dataset` option in the config file for the dataset
+  configuration. Supersedes `startup.dataset.config` which accepts a string
+  instead of a mapping [#1737]
+- Add alert to notify about duplicate sample/metric ingestion. [#1688]
 
 ### Changed
 - Reduced the verbosity of the logs emitted by the vacuum engine [#1715]